refactor($type-filter): prediction optimization
if a user were to append `type=user`, it would be accepted, but not converted to the proper value of
`users`. if type value is found in the accepted types array, the value will be updated to the
defined search type value rather than the user-entered value, in the case of minor user error.